reorganized SE3 tests
